design upgrade read me file added
