+++ client/src/components/table/addEntries.js
