+++ client/src/components/table/authGuard.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { onAuthStateChanged } from "firebase/auth";
-import { auth } from "../firebase/config";
+import { auth } from "../../firebase/config";
 import { Navigate } from "react-router-dom";
 
 const AuthGuard = ({ children }) => {
