+++ client/src/components/table/updateEntries.js
